move BaitType and BaitUpdateEvent to correct location
